app/requirements.txt
fastapi[standard]
sqlmodel
pymysql
cryptography
passlib
bcrypt<4.0.0
pyjwt
python-multipart
python-dotenv